+++ src/main/java/roomescape/dto/ReservationCreateVo.java
@@ -1,4 +1,4 @@
-package roomescape.vo;
+package roomescape.dto;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
